+++ CHANGELOG.md
@@ -4,6 +4,48 @@ Changelog
 `jira-wiki-markup` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+1.4.0
+-----
+
+Released 2021-05-25.
+
+* Allow quoted image parameters.
+
+* Added support for "smart links".
+
+* **API Change**: Add new constructors `SmartCard` and `SmartLink` to
+  Text.Jira.Markup.LinkType.
+
+1.3.5
+-----
+
+Released 2021-05-24.
+
+
+* Allow spaces and most unicode characters in attachment links.
+
+* No longer require a newline character after `{noformat}`.
+
+* Only allow URI path segment characters in bare links.
+
+* The `file:` schema is no longer allowed in bare links; these
+  rarely make sense.
+
+1.3.4
+-----
+
+Released 2021-03-13.
+
+* Fixed parsing of autolinks (i.e., of bare URLs in the text).
+  Previously an autolink would take up the rest of a line, as
+  spaces were allowed characters in these items.
+  
+* Emoji character sequences no longer cause parsing failures. This
+  was due to missing backtracking when emoji parsing fails.
+  
+* Block quotes are only rendered as `bq.` if they do not contain a
+  linebreak.
+
 1.3.3
 -----